Fixed backward compatibility of cluster space (and expansion) files
The code used to break when reading older ClusterExpansion files since
these files do not contain 'symprec' and 'position_tolerance'. The fix
simply sets this parameters to some default in order to be able to
(re)initialize the ClusterSpace. In border line cases where numerical
tolerance has been an issue before this could lead to inconsistencies
between ClusterSpace and the parameter vector stored in the
ClusterExpansion.

Note that I did not add a unittest for this fix but rather excluded
the respective line from the coverage since it is tedious to create
ClusterExpansion file in an outdated format "live" during testing.